Clean DuetPkg for ICC build.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@6836 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/DuetPkg/DxeIpl/X64/Paging.c
@@ -196,10 +196,10 @@ Description:
   VOID  *PageNumberBase;
 
   SizeOfMemorySpace -= EFI_2M_PAGE_BITS_NUM;
-  gPDEBitsNum        = MIN (SizeOfMemorySpace, EFI_MAX_ENTRY_BITS_NUM);
-  SizeOfMemorySpace  = SizeOfMemorySpace - gPDEBitsNum;
-  gPDPTEBitsNum      = MIN (SizeOfMemorySpace, EFI_MAX_ENTRY_BITS_NUM);
-  SizeOfMemorySpace  = SizeOfMemorySpace - gPDPTEBitsNum;
+  gPDEBitsNum        = (UINT8) MIN (SizeOfMemorySpace, EFI_MAX_ENTRY_BITS_NUM);
+  SizeOfMemorySpace  = (UINT8) (SizeOfMemorySpace - gPDEBitsNum);
+  gPDPTEBitsNum      = (UINT8) MIN (SizeOfMemorySpace, EFI_MAX_ENTRY_BITS_NUM);
+  SizeOfMemorySpace  = (UINT8) (SizeOfMemorySpace - gPDPTEBitsNum);
   gPML4BitsNum       = SizeOfMemorySpace;
   if (gPML4BitsNum > EFI_MAX_ENTRY_BITS_NUM) {
     return NULL;
